Use default font to draw checkmark
