+++ Emby.Drawing.ImageMagick/PlayedIndicatorDrawer.cs
@@ -2,11 +2,6 @@
 using MediaBrowser.Common.Configuration;
 using MediaBrowser.Common.Net;
 using MediaBrowser.Model.Drawing;
-using System;
-using System.IO;
-using System.Threading.Tasks;
-using MediaBrowser.Common.Progress;
-using MediaBrowser.Controller.IO;
 using MediaBrowser.Model.IO;
 
 namespace Emby.Drawing.ImageMagick
@@ -17,19 +12,19 @@ namespace Emby.Drawing.ImageMagick
         private const int OffsetFromTopRightCorner = 38;
 
         private readonly IApplicationPaths _appPaths;
-		private readonly IHttpClient _iHttpClient;
-		private readonly IFileSystem _fileSystem;
+        private readonly IHttpClient _iHttpClient;
+        private readonly IFileSystem _fileSystem;
 
-		public PlayedIndicatorDrawer(IApplicationPaths appPaths, IHttpClient iHttpClient, IFileSystem fileSystem)
+        public PlayedIndicatorDrawer(IApplicationPaths appPaths, IHttpClient iHttpClient, IFileSystem fileSystem)
         {
             _appPaths = appPaths;
             _iHttpClient = iHttpClient;
-			_fileSystem = fileSystem;
+            _fileSystem = fileSystem;
         }
 
-        public async Task DrawPlayedIndicator(MagickWand wand, ImageSize imageSize)
+        public void DrawPlayedIndicator(MagickWand wand, ImageSize imageSize)
         {
-            var x = imageSize.Width - OffsetFromTopRightCorner;
+            double x = imageSize.Width - OffsetFromTopRightCorner;
 
             using (var draw = new DrawingWand())
             {
@@ -43,83 +38,17 @@ namespace Emby.Drawing.ImageMagick
                     pixel.Opacity = 0;
                     pixel.Color = "white";
                     draw.FillColor = pixel;
-                    draw.Font = await DownloadFont("webdings.ttf", "https://github.com/MediaBrowser/Emby.Resources/raw/master/fonts/webdings.ttf", _appPaths, _iHttpClient, _fileSystem).ConfigureAwait(false);
                     draw.FontSize = FontSize;
                     draw.FontStyle = FontStyleType.NormalStyle;
                     draw.TextAlignment = TextAlignType.CenterAlign;
                     draw.FontWeight = FontWeightType.RegularStyle;
                     draw.TextAntialias = true;
-                    draw.DrawAnnotation(x + 4, OffsetFromTopRightCorner + 14, "a");
+                    draw.DrawAnnotation(x + 4, OffsetFromTopRightCorner + 14, "✓");
 
                     draw.FillColor = pixel;
                     wand.CurrentImage.DrawImage(draw);
                 }
             }
-        }
-
-		internal static string ExtractFont(string name, IApplicationPaths paths, IFileSystem fileSystem)
-        {
-            var filePath = Path.Combine(paths.ProgramDataPath, "fonts", name);
-
-			if (fileSystem.FileExists(filePath))
-            {
-                return filePath;
-            }
-
-            var namespacePath = typeof(PlayedIndicatorDrawer).Namespace + ".fonts." + name;
-            var tempPath = Path.Combine(paths.TempDirectory, Guid.NewGuid().ToString("N") + ".ttf");
-			fileSystem.CreateDirectory(fileSystem.GetDirectoryName(tempPath));
-
-            using (var stream = typeof(PlayedIndicatorDrawer).Assembly.GetManifestResourceStream(namespacePath))
-            {
-                using (var fileStream = new FileStream(tempPath, FileMode.Create, FileAccess.Write, FileShare.Read))
-                {
-                    stream.CopyTo(fileStream);
-                }
-            }
-
-			fileSystem.CreateDirectory(fileSystem.GetDirectoryName(filePath));
-
-            try
-            {
-				fileSystem.CopyFile(tempPath, filePath, false);
-            }
-            catch (IOException)
-            {
-
-            }
-
-            return tempPath;
-        }
-
-		internal static async Task<string> DownloadFont(string name, string url, IApplicationPaths paths, IHttpClient httpClient, IFileSystem fileSystem)
-        {
-            var filePath = Path.Combine(paths.ProgramDataPath, "fonts", name);
-
-			if (fileSystem.FileExists(filePath))
-            {
-                return filePath;
-            }
-
-            var tempPath = await httpClient.GetTempFile(new HttpRequestOptions
-            {
-                Url = url,
-                Progress = new SimpleProgress<double>()
-
-            }).ConfigureAwait(false);
-
-			fileSystem.CreateDirectory(fileSystem.GetDirectoryName(filePath));
-
-            try
-            {
-				fileSystem.CopyFile(tempPath, filePath, false);
-            }
-            catch (IOException)
-            {
-
-            }
-
-            return tempPath;
         }
     }
 }
